requirements-dev.txt
chromadb
sentence-transformers
datasets
numpy
tqdm
scikit-image
openai
python-dotenv
colormath==3.0.0
elkai==2.0.1
matplotlib